+++ ref-exchange/src/custom_keys.rs
@@ -11,6 +11,7 @@ pub const POOL_LIMIT: &str = "pl";
 // Key for client echo limit
 pub const CLIENT_ECHO_TOKEN_ID_WHITELIST: &str = "ce_tw";
 pub const CLIENT_ECHO_SENDER_ID_WHITELIST: &str = "ce_sw";
+pub const SECURE_SENDER_WHITELIST: &str = "ce_ssw";
 
 // Key for swap volume
 pub const SWAP_VOLUME_KEY: &str = "sv";