+++ ref-exchange/release_notes.md
@@ -1,5 +1,12 @@
 # Release Notes
 
+### Version 1.9.14
+```
+G1mrfT8dceTrrjn95LeCq6HS2LbFktd9XWp49oXZAdXH
+```
+1. add secure_sender_whitelist in client echo feature, if a sender falls in ssw, client_echo_token whitelist would be ignored.
+2. add prefix wildcard '\*' support in client_echo whitelists. In practise, it always starts with '\*.' to indicate sub-accounts.
+
 ### Version 1.9.13
 ```
 4izBbspd1Uiu1vjYcWpkebzoVxqmiZK7ZXSn4b2fAF2D